mcp_server/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.1.0
asyncpg>=0.29.0
python-dateutil>=2.8.0
fastmcp>=2.0.0
mcp>=1.0.0
python-multipart>=0.0.7
python-dotenv>=1.0.0
PyJWT[crypto]>=2.8.0

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
